Доделано.
Функционал доделан
