+++ Resources/Locale/ru-RU/Imperial/EverythingStamp/EverythingStamp.ftl
@@ -1,0 +1,8 @@
+everything-stamp-chosen-stamp = Выбрана печать
+everything-stamp-new-stamp-added = Печать успешно скопирована
+everything-stamp-new-stamp-already-added = Эта печать уже была скопирована
+uplink-everything-stamp-name = Печать агента синдиката
+uplink-everything-stamp-desc = Печать, применяющаяся различными спецслужбами, для получения копий печатей и дальнейшей подделки документов. Поставляется с копией печати синдиката.
+ent-EverythingStamp = Печать агента синдиката
+    .desc = Странная печать.
+
