+++ Resources/Prototypes/Entities/Objects/Misc/rubber_stamp.yml
@@ -271,16 +271,3 @@
     stampState: "paper_stamp-detective"
   - type: Sprite
     state: stamp-detective
-
-- type: entity
-  name: everything stamp
-  parent: BaseItem
-  id: everythingStamp
-  components:
-  - type: Stamp
-  - type: EverythingStamp
-  - type: Sprite
-    sprite: Objects/Misc/stamps.rsi
-    state: stamp-mime
-  - type: Item
-    size: Tiny
